requirements.txt
fitz==0.0.1.dev2
numpy==2.2.2
opencv_python==4.11.0.86
Pillow==11.1.0
pytesseract==0.3.13
streamlit==1.41.1
